Removing AllowsCultureVariation from ContentController, so we can get rid of OverrideAuthorization attribute
